issue: 4367367 Rename XLIO Socket API to Ultra API
