+++ src/core/dev/ring.h
@@ -203,7 +203,7 @@ public:
     struct tcp_seg *get_tcp_segs(uint32_t num);
     void put_tcp_segs(struct tcp_seg *seg);
 
-    // XLIO ZC API
+    // XLIO Ultra API
     void set_poll_group(poll_group *p_group) { m_p_group = p_group; }
     poll_group *get_poll_group() const { return m_p_group; }
 
@@ -218,7 +218,7 @@ protected:
     uint32_t m_tcp_seg_count = 0U;
     lock_spin_recursive m_tcp_seg_lock;
 
-    // XLIO ZC API
+    // XLIO Ultra API
     poll_group *m_p_group = nullptr;
 
     int m_if_index = 0; /* Interface index */
