+++ tests/extra_api/xlio_ultra_api.c
@@ -4,9 +4,9 @@
  * SPDX-License-Identifier: GPL-2.0-only or BSD-2-Clause
  */
 
-/* g++ -I./install/include -L./install/lib -L../dpcp/install/lib -o test xlio_socket_api_listen.c -lxlio -lm -lnl-3 -ldpcp -libverbs -lmlx5 -lrdmacm -lnl-route-3 -g3 */
+/* g++ -I./install/include -L./install/lib -L../dpcp/install/lib -o test xlio_ultra_api.c -lxlio -lm -lnl-3 -ldpcp -libverbs -lmlx5 -lrdmacm -lnl-route-3 -g3 */
 /* LD_LIBRARY_PATH=./install/lib:../dpcp/install/lib ./test */
-/* Use `nc <IP> <PORT>` on the remote side */
+/* Use `nc -l 8080` on the remote side */
 
 #include <assert.h>
 #include <stdbool.h>
@@ -20,22 +20,15 @@
 #include <mellanox/xlio.h>
 #include <infiniband/verbs.h>
 
-#define LISTEN_MAGIC 0xdeadc0de
+#define TEST_USERDATA_MAGIC 0xfeedbeef
+#define FAKE_PORT           65535
 
 static bool quit = false;
 static bool terminated = false;
+static int g_test_events;
 static int g_comp_events = 0;
 static char sndbuf[256];
-
-static struct ibv_pd *pd = NULL;
 static struct ibv_mr *mr_buf;
-
-static unsigned sock_nr = 0;
-static unsigned destroyed_sock_nr = 0;
-static xlio_socket_t sock_arr[32];
-
-static unsigned short listen_port = 8080;
-static const char *listen_ip = "";
 
 static void memory_cb(void *data, size_t size, size_t page_size)
 {
@@ -77,11 +70,7 @@ static void socket_event_cb(xlio_socket_t sock, uintptr_t userdata_sq, int event
         printf("Connection closed passively (sock=%lx).\n", userdata_sq);
     } else if (event == XLIO_SOCKET_EVENT_TERMINATED) {
         printf("Connection terminated (sock=%lx).\n", userdata_sq);
-        if (userdata_sq == LISTEN_MAGIC) {
-            terminated = true;
-        } else {
-            ++destroyed_sock_nr;
-        }
+        terminated = true;
     } else {
         printf("Event callback: event=%d value=%d (sock=%lx).\n", event, value, userdata_sq);
         if (event == XLIO_SOCKET_EVENT_ERROR) {
@@ -93,6 +82,7 @@ static void socket_event_cb(xlio_socket_t sock, uintptr_t userdata_sq, int event
 static void socket_comp_cb(xlio_socket_t sock, uintptr_t userdata_sq, uintptr_t userdata_op)
 {
     const char *reply_msg = "completed\n";
+    const char *inline_msg = "inline\n";
 
     printf("Completed zcopy buffer userdata_sq=%lx userdata_op=%lx.\n", userdata_sq, userdata_op);
     assert(userdata_sq != 0);
@@ -105,6 +95,7 @@ static void socket_comp_cb(xlio_socket_t sock, uintptr_t userdata_sq, uintptr_t 
          * XLIO_SOCKET_EVENT_TERMINATED event arrives.
          */
         send_single_msg(sock, reply_msg, strlen(reply_msg), 0, 0);
+        send_inline_msg(sock, inline_msg, strlen(inline_msg), 0, 0);
     }
 }
 
@@ -127,50 +118,122 @@ static void socket_rx_cb(xlio_socket_t sock, uintptr_t userdata_sq, void *data, 
     xlio_socket_buf_free(sock, buf);
 }
 
-static void socket_accept_cb(xlio_socket_t sock, xlio_socket_t parent_sock,
-                             uintptr_t parent_userdata)
+static void test_event_cb(xlio_socket_t sock, uintptr_t userdata_sq, int event, int value)
 {
-    struct sockaddr_in6 sa = {};
-    socklen_t len = sizeof(sa);
+    (void)sock;
+    (void)value;
+    assert(userdata_sq = TEST_USERDATA_MAGIC);
+
+    printf("Test event callback: event=%d value=%d.\n", event, value);
+
+    if (event == XLIO_SOCKET_EVENT_ERROR || event == XLIO_SOCKET_EVENT_TERMINATED) {
+        ++g_test_events;
+    }
+}
+
+static void test_comp_cb(xlio_socket_t sock, uintptr_t userdata_sq, uintptr_t userdata_op)
+{
+    (void)sock;
+    (void)userdata_op;
+    assert(userdata_sq = TEST_USERDATA_MAGIC);
+}
+
+static void test_rx_cb(xlio_socket_t sock, uintptr_t userdata_sq, void *data, size_t len,
+                       struct xlio_buf *buf)
+{
+    (void)data;
+    (void)len;
+    assert(userdata_sq = TEST_USERDATA_MAGIC);
+    xlio_socket_buf_free(sock, buf);
+}
+
+static void test_multi_groups(const char *ip)
+{
+    xlio_poll_group_t group1;
+    xlio_poll_group_t group2;
+    xlio_poll_group_t group3;
+    xlio_socket_t sock1_1;
+    xlio_socket_t sock1_2;
+    xlio_socket_t sock2;
+    xlio_socket_t sock3;
     int rc;
-    unsigned short port = 0;
-    char buf[64] = "";
 
-    rc = xlio_socket_getpeername(sock, (struct sockaddr *)&sa, &len);
-    if (rc != 0) {
-        printf("Failed to get peername of an incoming socket.\n");
-        // XXX Cannot close the socket in this callback in the current implementation.
-        return;
+    struct xlio_poll_group_attr gattr = {
+        .socket_event_cb = &test_event_cb,
+        .socket_comp_cb = &test_comp_cb,
+        .socket_rx_cb = &test_rx_cb,
+    };
+
+    rc = xlio_poll_group_create(&gattr, &group1);
+    assert(rc == 0);
+    rc = xlio_poll_group_create(&gattr, &group2);
+    assert(rc == 0);
+
+    gattr.flags = XLIO_GROUP_FLAG_SAFE;
+    rc = xlio_poll_group_create(&gattr, &group3);
+    assert(rc == 0);
+
+    struct xlio_socket_attr sattr = {
+        .domain = AF_INET,
+        .userdata_sq = TEST_USERDATA_MAGIC,
+    };
+
+    sattr.group = group1;
+    rc = xlio_socket_create(&sattr, &sock1_1);
+    assert(rc == 0);
+    rc = xlio_socket_create(&sattr, &sock1_2);
+    assert(rc == 0);
+    sattr.group = group2;
+    rc = xlio_socket_create(&sattr, &sock2);
+    assert(rc == 0);
+    sattr.group = group3;
+    rc = xlio_socket_create(&sattr, &sock3);
+    assert(rc == 0);
+
+    struct sockaddr_in addr = {};
+    addr.sin_family = AF_INET;
+    addr.sin_port = htons(FAKE_PORT);
+    rc = inet_aton(ip, &addr.sin_addr);
+    assert(rc != 0);
+
+    g_test_events = 0;
+    /* Connect will fail, we need it to allocate rings for the checks below. */
+    rc = xlio_socket_connect(sock1_1, (struct sockaddr *)&addr, sizeof(addr));
+    assert(rc == 0);
+    rc = xlio_socket_connect(sock1_2, (struct sockaddr *)&addr, sizeof(addr));
+    assert(rc == 0);
+    rc = xlio_socket_connect(sock2, (struct sockaddr *)&addr, sizeof(addr));
+    assert(rc == 0);
+    rc = xlio_socket_connect(sock3, (struct sockaddr *)&addr, sizeof(addr));
+    assert(rc == 0);
+
+    /* TODO There is no API to check expected internal ring distribution. */
+
+    /* Wait for ERROR events (ECONREFUSED). */
+    while (g_test_events < 4) {
+        xlio_poll_group_poll(group1);
+        xlio_poll_group_poll(group2);
+        xlio_poll_group_poll(group3);
     }
 
-    switch (sa.sin6_family) {
-    case AF_INET:
-        inet_ntop(AF_INET, &((struct sockaddr_in *)&sa)->sin_addr, buf, sizeof(buf));
-        port = ntohs(((struct sockaddr_in *)&sa)->sin_port);
-        break;
-    case AF_INET6:
-        inet_ntop(AF_INET6, &sa.sin6_addr, buf, sizeof(buf));
-        port = ntohs(sa.sin6_port);
-        break;
-    default:
-        printf("Unknown AF: %u.\n", sa.sin6_family);
+    g_test_events = 0;
+    xlio_socket_destroy(sock1_1);
+    xlio_socket_destroy(sock1_2);
+    xlio_socket_destroy(sock2);
+    xlio_socket_destroy(sock3);
+
+    /* Wait for TERMINATED events. */
+    while (g_test_events < 4) {
+        xlio_poll_group_poll(group1);
+        xlio_poll_group_poll(group2);
+        xlio_poll_group_poll(group3);
     }
 
-    if (!pd) {
-        pd = xlio_socket_get_pd(sock);
-        assert(pd != NULL);
-        mr_buf = ibv_reg_mr(pd, sndbuf, sizeof(sndbuf), IBV_ACCESS_LOCAL_WRITE);
-        assert(mr_buf != NULL);
-    }
+    xlio_poll_group_destroy(group1);
+    xlio_poll_group_destroy(group2);
+    xlio_poll_group_destroy(group3);
 
-    sock_arr[sock_nr++] = sock;
-
-    uintptr_t userdata_sq = (uintptr_t)sock_nr;
-    rc = xlio_socket_update(sock, 0, userdata_sq);
-    if (rc != 0) {
-        printf("Failed to update context of an incoming socket.\n");
-    }
-    printf("Accepted incoming connection from %s:%u userdata_sq=%lx.\n", buf, port, userdata_sq);
+    printf("Multi group test done.\n");
 }
 
 int main(int argc, char **argv)
@@ -187,19 +250,20 @@ int main(int argc, char **argv)
         .socket_event_cb = &socket_event_cb,
         .socket_comp_cb = &socket_comp_cb,
         .socket_rx_cb = &socket_rx_cb,
-        .socket_accept_cb = &socket_accept_cb,
     };
 
-    if (argc > 1 && (strcmp(argv[1], "--help") == 0 || strcmp(argv[1], "-h") == 0)) {
-        printf("Usage: %s [IP [PORT]]\n", argv[0]);
-        printf("Run 'nc <IP> <PORT>' on the client side.\n");
+    if (argc < 2) {
+        printf("Usage: %s <IP>\n", argv[0]);
+        printf("Run 'nc -l 8080' on the server with the <IP> address.\n");
         printf("Type messages on the nc side.\n");
-        printf("Message 'quit' or 'exit' will terminate the server.\n");
-        return 0;
+        printf("Message 'quit' or 'exit' will terminate the client.\n");
+        return 1;
     }
 
     rc = xlio_init_ex(&iattr);
     assert(rc == 0);
+
+    test_multi_groups(argv[1]);
 
     rc = xlio_poll_group_create(&gattr, &group);
     assert(rc == 0);
@@ -209,36 +273,28 @@ int main(int argc, char **argv)
     struct xlio_socket_attr sattr = {
         .domain = AF_INET,
         .group = group,
-        .userdata_sq = LISTEN_MAGIC,
+        .userdata_sq = 0xdeadc0de,
     };
 
     rc = xlio_socket_create(&sattr, &sock);
     assert(rc == 0);
 
-    printf("Listen socket created.\n");
+    printf("Socket created, connecting to %s:8080.\n", argv[1]);
 
-    if (argc > 1) {
-        listen_ip = argv[1];
-    }
-    if (argc > 2) {
-        listen_port = atoi(argv[2]);
-    }
     struct sockaddr_in addr = {};
     addr.sin_family = AF_INET;
-    addr.sin_port = htons(listen_port);
-    if (listen_ip[0] != '\0') {
-        rc = inet_aton(argv[1], &addr.sin_addr);
-        assert(rc != 0);
-    }
-    rc = xlio_socket_bind(sock, (struct sockaddr *)&addr, sizeof(addr));
+    addr.sin_port = htons(8080);
+    rc = inet_aton(argv[1], &addr.sin_addr);
+    assert(rc != 0);
+
+    rc = xlio_socket_connect(sock, (struct sockaddr *)&addr, sizeof(addr));
     assert(rc == 0);
 
-    printf("Listen socket bound to %s:%u.\n", listen_ip, listen_port);
+    struct ibv_pd *pd = xlio_socket_get_pd(sock);
+    assert(pd != NULL);
+    mr_buf = ibv_reg_mr(pd, sndbuf, sizeof(sndbuf), IBV_ACCESS_LOCAL_WRITE);
+    assert(mr_buf != NULL);
 
-    rc = xlio_socket_listen(sock);
-    assert(rc == 0);
-
-    printf("Listen.\n");
     printf("Starting polling loop.\n");
 
     while (!quit) {
@@ -249,24 +305,17 @@ int main(int argc, char **argv)
 
     rc = xlio_socket_destroy(sock);
     assert(rc == 0);
-    for (unsigned i = 0; i < sock_nr; ++i) {
-        rc = xlio_socket_destroy(sock_arr[i]);
-        assert(rc == 0);
-    }
-    while (!terminated || destroyed_sock_nr < sock_nr) {
+
+    while (!terminated) {
         xlio_poll_group_poll(group);
     }
-
-    printf("All the sockets are destroyed.\n");
 
     rc = xlio_poll_group_destroy(group);
     assert(rc == 0);
 
     printf("Zerocopy completion events: %d\n", g_comp_events);
 
-    if (mr_buf) {
-        ibv_dereg_mr(mr_buf);
-    }
+    ibv_dereg_mr(mr_buf);
     xlio_exit();
 
     return 0;
