+++ tests/gtest/xlio_ultra_api/xlio_socket_listen_connect.cc
@@ -8,28 +8,20 @@
 #include "common/log.h"
 #include "common/sys.h"
 #include "common/base.h"
-#include <infiniband/verbs.h>
-#include <pthread.h>
-#include <unistd.h>
 #include "core/xlio_base.h"
 
 #if defined(EXTRA_API_ENABLED) && (EXTRA_API_ENABLED == 1)
 
 static int connected_counter = 0;
 static int terminated_counter = 0;
-static int rx_cb_counter = 0;
-static int comp_cb_counter = 0;
-static const char *data_to_send = "I Love XLIO!";
-static struct ibv_pd *pd = NULL;
-static struct ibv_mr *mr_buf;
-static char sndbuf[256];
 static std::vector<xlio_socket_t> accepted_sockets;
 
-class zc_api_xlio_socket_send_receive_2 : public xlio_zc_api_base {
+class xlio_ultra_api_socket_listen_connect : public xlio_ultra_api_base {
 public:
     virtual void SetUp() { errno = EOK; };
     virtual void TearDown() {};
     void destroy_poll_group(xlio_poll_group_t group) { base_destroy_poll_group(group); }
+
     static void socket_event_cb(xlio_socket_t sock, uintptr_t userdata_sq, int event, int value)
     {
         UNREFERENCED_PARAMETER(sock);
@@ -48,7 +40,6 @@ public:
         UNREFERENCED_PARAMETER(sock);
         UNREFERENCED_PARAMETER(userdata_sq);
         UNREFERENCED_PARAMETER(userdata_op);
-        comp_cb_counter++;
     }
 
     static void socket_rx_cb(xlio_socket_t sock, uintptr_t userdata_sq, void *data, size_t len,
@@ -56,13 +47,9 @@ public:
     {
         UNREFERENCED_PARAMETER(sock);
         UNREFERENCED_PARAMETER(userdata_sq);
-        rx_cb_counter++;
-
-        // Assume that the data_to_send is received in one packet
-        if (memcmp(data, data_to_send, len) != 0) {
-            GTEST_FAIL();
-        }
-        xlio_api->xlio_socket_buf_free(sock, buf);
+        UNREFERENCED_PARAMETER(data);
+        UNREFERENCED_PARAMETER(len);
+        UNREFERENCED_PARAMETER(buf);
     }
 
     static void socket_accept_cb(xlio_socket_t sock, xlio_socket_t parent_sock,
@@ -71,29 +58,22 @@ public:
         UNREFERENCED_PARAMETER(sock);
         UNREFERENCED_PARAMETER(parent_sock);
         UNREFERENCED_PARAMETER(parent_userdata);
-        int rc = xlio_api->xlio_socket_update(sock, 0, 0x1);
-        ASSERT_EQ(rc, 0);
         accepted_sockets.push_back(sock);
         connected_counter++;
-        pd = xlio_api->xlio_socket_get_pd(sock);
-        ASSERT_TRUE(pd != NULL);
-        mr_buf = ibv_reg_mr(pd, sndbuf, sizeof(sndbuf), IBV_ACCESS_LOCAL_WRITE);
-        ASSERT_TRUE(mr_buf != NULL);
-        base_send_single_msg(sock, data_to_send, strlen(data_to_send), 0x1, 0, mr_buf, sndbuf);
     }
 };
 
 /**
- * @test zc_api_xlio_socket_send_receive_2.ti_1
+ * @test socket_connect.ti_1
  * @brief
- *    Create TCP socket/connect/send(target)/receive(initiator)
+ *    Create TCP socket/listen(target)/connect(initiator)
  * @details
  */
-TEST_F(zc_api_xlio_socket_send_receive_2, ti_1)
+TEST_F(xlio_ultra_api_socket_listen_connect, ti_1)
 {
     int rc;
     int pid = fork();
-    xlio_zc_api_base::SetUp();
+    xlio_ultra_api_base::SetUp();
     xlio_poll_group_t group;
     xlio_socket_t sock;
 
@@ -105,51 +85,47 @@ TEST_F(zc_api_xlio_socket_send_receive_2, ti_1)
         .group = group,
         .userdata_sq = 0,
     };
-    base_create_socket(&sattr, &sock);
+
     if (pid == 0) {
         // Child process - server side
+        base_create_socket(&sattr, &sock);
         rc = xlio_api->xlio_socket_bind(sock, (struct sockaddr *)&server_addr, sizeof(server_addr));
         ASSERT_EQ(0, rc);
-
         rc = xlio_api->xlio_socket_listen(sock);
         ASSERT_EQ(0, rc);
-
-        barrier_fork(pid, true);
-
-        while (connected_counter < 1 || comp_cb_counter < 1) {
+        barrier_fork(pid, true); // Tell parent that we are listening
+        while (connected_counter < 1) {
             xlio_api->xlio_poll_group_poll(group);
         }
-
         base_wait_for_delayed_acks(group);
-
-        barrier_fork(pid, true);
-
+        barrier_fork(pid, true); // Tell parent that we got last ack
         base_destroy_socket(sock);
         base_cleanup_accepted_sockets(accepted_sockets);
         while (terminated_counter < 1) {
             xlio_api->xlio_poll_group_poll(group);
         }
-
         destroy_poll_group(group);
         exit(testing::Test::HasFailure());
     } else {
         // Parent process - client side
+        base_create_socket(&sattr, &sock);
+
         rc = xlio_api->xlio_socket_bind(sock, (struct sockaddr *)&client_addr, sizeof(client_addr));
         ASSERT_EQ(0, rc);
 
-        barrier_fork(pid, true); // Wait for child to bind and listen
+        barrier_fork(pid, true); // Wait for child to listen
 
         rc = xlio_api->xlio_socket_connect(sock, (struct sockaddr *)&server_addr,
                                            sizeof(server_addr));
         ASSERT_EQ(0, rc);
 
-        while (connected_counter < 1 || rx_cb_counter < 1) {
+        while (connected_counter < 1) {
             xlio_api->xlio_poll_group_poll(group);
         }
 
         base_wait_for_delayed_acks(group);
 
-        barrier_fork(pid, true); // Wait for child to accept + receive last ack
+        barrier_fork(pid, true); // Wait for child to get last ack
 
         base_destroy_socket(sock);
         while (terminated_counter < 1) {
