+++ tests/gtest/xlio_ultra_api/xlio_socket_send_receive_2.cc
@@ -20,29 +20,22 @@ static int terminated_counter = 0;
 static int rx_cb_counter = 0;
 static int comp_cb_counter = 0;
 static const char *data_to_send = "I Love XLIO!";
-static int data_bytes_to_be_sent = strlen(data_to_send) * 1000;
-static int data_sent_completed = 0;
-static int data_received = 0;
 static struct ibv_pd *pd = NULL;
 static struct ibv_mr *mr_buf;
 static char sndbuf[256];
-static bool do_migrate = false;
 static std::vector<xlio_socket_t> accepted_sockets;
 
-class zc_api_xlio_socket_migrate : public xlio_zc_api_base {
+class xlio_ultra_api_socket_send_receive_2 : public xlio_ultra_api_base {
 public:
     virtual void SetUp() { errno = EOK; };
     virtual void TearDown() {};
     void destroy_poll_group(xlio_poll_group_t group) { base_destroy_poll_group(group); }
     static void socket_event_cb(xlio_socket_t sock, uintptr_t userdata_sq, int event, int value)
     {
+        UNREFERENCED_PARAMETER(sock);
         UNREFERENCED_PARAMETER(userdata_sq);
         UNREFERENCED_PARAMETER(value);
         if (event == XLIO_SOCKET_EVENT_ESTABLISHED) {
-            pd = xlio_api->xlio_socket_get_pd(sock);
-            ASSERT_TRUE(pd != NULL);
-            mr_buf = ibv_reg_mr(pd, sndbuf, sizeof(sndbuf), IBV_ACCESS_LOCAL_WRITE);
-            ASSERT_TRUE(mr_buf != NULL);
             connected_counter++;
         } else if (event == XLIO_SOCKET_EVENT_CLOSED) {
             terminated_counter++;
@@ -54,46 +47,53 @@ public:
     {
         UNREFERENCED_PARAMETER(sock);
         UNREFERENCED_PARAMETER(userdata_sq);
+        UNREFERENCED_PARAMETER(userdata_op);
         comp_cb_counter++;
-        data_sent_completed += userdata_op;
     }
 
     static void socket_rx_cb(xlio_socket_t sock, uintptr_t userdata_sq, void *data, size_t len,
                              struct xlio_buf *buf)
     {
+        UNREFERENCED_PARAMETER(sock);
         UNREFERENCED_PARAMETER(userdata_sq);
-        UNREFERENCED_PARAMETER(data);
-        if (rx_cb_counter == 0) {
-            do_migrate = true;
-        }
         rx_cb_counter++;
-        data_received += len;
+
+        // Assume that the data_to_send is received in one packet
+        if (memcmp(data, data_to_send, len) != 0) {
+            GTEST_FAIL();
+        }
         xlio_api->xlio_socket_buf_free(sock, buf);
     }
 
     static void socket_accept_cb(xlio_socket_t sock, xlio_socket_t parent_sock,
                                  uintptr_t parent_userdata)
     {
+        UNREFERENCED_PARAMETER(sock);
         UNREFERENCED_PARAMETER(parent_sock);
         UNREFERENCED_PARAMETER(parent_userdata);
         int rc = xlio_api->xlio_socket_update(sock, 0, 0x1);
         ASSERT_EQ(rc, 0);
         accepted_sockets.push_back(sock);
         connected_counter++;
+        pd = xlio_api->xlio_socket_get_pd(sock);
+        ASSERT_TRUE(pd != NULL);
+        mr_buf = ibv_reg_mr(pd, sndbuf, sizeof(sndbuf), IBV_ACCESS_LOCAL_WRITE);
+        ASSERT_TRUE(mr_buf != NULL);
+        base_send_single_msg(sock, data_to_send, strlen(data_to_send), 0x1, 0, mr_buf, sndbuf);
     }
 };
 
 /**
- * @test zc_api_xlio_socket_migrate.ti_1
+ * @test xlio_ultra_api_socket_send_receive_2.ti_1
  * @brief
- *    Create TCP socket/connect/send(initiator)/receive(target)
+ *    Create TCP socket/connect/send(target)/receive(initiator)
  * @details
  */
-TEST_F(zc_api_xlio_socket_migrate, ti_1)
+TEST_F(xlio_ultra_api_socket_send_receive_2, ti_1)
 {
     int rc;
     int pid = fork();
-    xlio_zc_api_base::SetUp();
+    xlio_ultra_api_base::SetUp();
     xlio_poll_group_t group;
     xlio_socket_t sock;
 
@@ -105,13 +105,9 @@ TEST_F(zc_api_xlio_socket_migrate, ti_1)
         .group = group,
         .userdata_sq = 0,
     };
+    base_create_socket(&sattr, &sock);
     if (pid == 0) {
-        xlio_poll_group_t group_2;
-        base_create_poll_group(&group_2, &socket_event_cb, &socket_comp_cb, &socket_rx_cb,
-                               &socket_accept_cb);
-
-        base_create_socket(&sattr, &sock);
-
+        // Child process - server side
         rc = xlio_api->xlio_socket_bind(sock, (struct sockaddr *)&server_addr, sizeof(server_addr));
         ASSERT_EQ(0, rc);
 
@@ -120,19 +116,11 @@ TEST_F(zc_api_xlio_socket_migrate, ti_1)
 
         barrier_fork(pid, true);
 
-        while (data_received < data_bytes_to_be_sent) {
+        while (connected_counter < 1 || comp_cb_counter < 1) {
             xlio_api->xlio_poll_group_poll(group);
-            xlio_api->xlio_poll_group_poll(group_2);
-            if (do_migrate) {
-                xlio_api->xlio_socket_detach_group(accepted_sockets.back());
-                xlio_api->xlio_socket_attach_group(accepted_sockets.back(), group_2);
-                do_migrate = false;
-            }
         }
 
         base_wait_for_delayed_acks(group);
-
-        ASSERT_EQ(data_received, data_bytes_to_be_sent);
 
         barrier_fork(pid, true);
 
@@ -145,32 +133,23 @@ TEST_F(zc_api_xlio_socket_migrate, ti_1)
         destroy_poll_group(group);
         exit(testing::Test::HasFailure());
     } else {
-        base_create_socket(&sattr, &sock);
-
+        // Parent process - client side
         rc = xlio_api->xlio_socket_bind(sock, (struct sockaddr *)&client_addr, sizeof(client_addr));
         ASSERT_EQ(0, rc);
 
-        barrier_fork(pid, true); // wait for child to bind and listen
+        barrier_fork(pid, true); // Wait for child to bind and listen
 
         rc = xlio_api->xlio_socket_connect(sock, (struct sockaddr *)&server_addr,
                                            sizeof(server_addr));
         ASSERT_EQ(0, rc);
 
-        int sent_bytes = 0;
-        while (data_sent_completed < data_bytes_to_be_sent) {
+        while (connected_counter < 1 || rx_cb_counter < 1) {
             xlio_api->xlio_poll_group_poll(group);
-            if (connected_counter > 0 && sent_bytes < data_bytes_to_be_sent) {
-                base_send_single_msg(sock, data_to_send, strlen(data_to_send), strlen(data_to_send),
-                                     0, mr_buf, sndbuf);
-                sent_bytes += strlen(data_to_send);
-            }
         }
 
         base_wait_for_delayed_acks(group);
 
-        ASSERT_EQ(data_sent_completed, data_bytes_to_be_sent);
-
-        barrier_fork(pid, true); // wait for child to accept + receive last ack
+        barrier_fork(pid, true); // Wait for child to accept + receive last ack
 
         base_destroy_socket(sock);
         while (terminated_counter < 1) {
